+++ src/Extension.php
@@ -8,14 +8,10 @@
 namespace Studiometa\TwigToolkit;
 
 use Studiometa\TwigToolkit\Helpers\Html;
-use Studiometa\TwigToolkit\Helpers\Template;
 use Studiometa\TwigToolkit\Helpers\Url;
-use Studiometa\TwigToolkit\Node\ElementNode;
 use Studiometa\TwigToolkit\TokenParser\ElementTokenParser;
 use Twig\Extension\AbstractExtension;
-use Twig\Loader\FilesystemLoader;
 use Twig\TokenParser\TokenParserInterface;
-use Twig\TwigFilter;
 use Twig\TwigFunction;
 
 /**
@@ -26,18 +22,6 @@ use Twig\TwigFunction;
  */
 class Extension extends AbstractExtension
 {
-    /**
-     * Register the `@meta` namespace if the $loader parameter is specifier.
-     *
-     * @param FilesystemLoader|null $loader The Twig FilesystemLoader instance.
-     */
-    public function __construct(FilesystemLoader $loader = null)
-    {
-        if ($loader) {
-            Template::addMetaNamespace($loader);
-        }
-    }
-
     /**
      * Returns the token parser instances to add to the existing list.
      *
@@ -58,15 +42,6 @@ class Extension extends AbstractExtension
     public function getFunctions()
     {
         return [
-            /** @deprecated 1.0.1 Use the `html_classes` function instead. */
-            new TwigFunction('class', [Html::class, 'renderClass']),
-            /** @deprecated 1.0.1 Use the `html_attributrs` function instead. */
-            new TwigFunction(
-                'attributes',
-                [Html::class, 'renderAttributes'],
-                ['needs_environment' => true, 'is_safe' => ['html']]
-            ),
-
             new TwigFunction('html_classes', [Html::class, 'renderClass']),
             new TwigFunction('html_styles', [Html::class, 'renderStyleAttribute']),
             new TwigFunction(
@@ -82,18 +57,6 @@ class Extension extends AbstractExtension
                 'twig_toolkit_url',
                 [Url::class, 'fromString'],
             ),
-        ];
-    }
-
-    /**
-     * Return a list of filters.
-     *
-     * @return TwigFilter[]
-     */
-    public function getFilters()
-    {
-        return [
-            new TwigFilter('merge_html_attributes', [Html::class, 'mergeAttributes']),
         ];
     }
 }