+++ src/Node/ElementNode.php
@@ -10,6 +10,8 @@ namespace Studiometa\TwigToolkit\Node;
 use Twig\Compiler;
 use Twig\Node\Node;
 use Twig\Node\NodeCaptureInterface;
+use Twig\Attribute\YieldReady;
+use Twig\Node\Expression\AbstractExpression;
 
 /**
  * Class ElementNode
@@ -17,30 +19,47 @@ use Twig\Node\NodeCaptureInterface;
  * @author Studio Meta <agence@studiometa.fr>
  * @since 1.0.0
  */
+#[YieldReady]
 class ElementNode extends Node implements NodeCaptureInterface
 {
+    public function __construct(
+        AbstractExpression $name,
+        Node $body,
+        ?AbstractExpression $variables,
+        int $lineno,
+    ) {
+        $nodes = [
+            'name' => $name,
+            'body' => $body,
+        ];
+        if (null !== $variables) {
+            $nodes['variables'] = $variables;
+        }
+
+        $capture = $body->count() > 0;
+
+        if ($body->hasAttribute('data')) {
+            $data = $body->getAttribute('data');
+            $capture = is_string($data) ? !empty(trim($data)) : $capture;
+        }
+
+        parent::__construct($nodes, ['capture' => $capture], $lineno);
+    }
     /**
      * @inheritdoc
      */
     public function compile(Compiler $compiler)
     {
+        $compiler->addDebugInfo($this);
+
         if ($this->getAttribute('capture')) {
-            if (class_exists('Twig\Node\CaptureNode')) {
-                $compiler->write('$body = ');
-                $node = new \Twig\Node\CaptureNode(
-                    $this->getNode('body'),
-                    $this->getNode('body')->lineno,
-                    $this->getNode('body')->tag
-                );
-                $node->setAttribute('with_blocks', true);
-                $compiler->subcompile($node);
-            } else {
-                $compiler->write("ob_start();\n");
-                $compiler->subcompile($this->getNode('body'));
-                $compiler->write(
-                    '$body = ("" === $tmp = ob_get_clean()) ? null : new Markup($tmp, $this->env->getCharset());'
-                );
-            }
+            $compiler->write('$body = ');
+            $node = new \Twig\Node\CaptureNode(
+                $this->getNode('body'),
+                $this->getNode('body')->getTemplateLine()
+            );
+            $node->setAttribute('with_blocks', true);
+            $compiler->subcompile($node);
             $compiler->raw("\n");
         } else {
             $compiler->write('$body = null;')->raw("\n");
@@ -49,12 +68,12 @@ class ElementNode extends Node implements NodeCaptureInterface
         // Element
         $compiler
             ->addDebugInfo($this)
-            ->write('echo $dom = \Studiometa\TwigToolkit\Helpers\Html::renderTag($this->env, ')
-            ->subcompile($this->getNode('element'))
+            ->write('yield $dom = \Studiometa\TwigToolkit\Helpers\Html::renderTag($this->env, ')
+            ->subcompile($this->getNode('name'))
             ->raw(', ');
 
-        if ($this->hasNode('attrs')) {
-            $compiler->subcompile($this->getNode('attrs'));
+        if ($this->hasNode('variables')) {
+            $compiler->subcompile($this->getNode('variables'));
         } else {
             $compiler->raw('[]');
         }